combine other actors into misc.py
